+++ pyramid_google_login/views.py
@@ -1,11 +1,12 @@
 import logging
 
 from pyramid.view import view_config
+# from pyramid.settings import asbool
 from pyramid.security import (remember, forget, NO_PERMISSION_REQUIRED)
 from pyramid.httpexceptions import HTTPFound
 
 from pyramid_google_login import SETTINGS_PREFIX
-from pyramid_google_login import redirect_to_signin
+from pyramid_google_login import redirect_to_signin, find_landing_path
 from pyramid_google_login import AuthFailed
 from pyramid_google_login.google_oauth2 import (build_authorize_url,
                                                 exchange_token_from_code,
@@ -24,18 +25,27 @@ log = logging.getLogger(__name__)
              renderer='pyramid_google_login:templates/signin.mako')
 def signin(request):
     settings = request.registry.settings
+    # auto_redirect = asbool(settings.get(SETTINGS_PREFIX + 'auto_redirect',
+    #                                     True))
     signin_banner = settings.get(SETTINGS_PREFIX + 'signin_banner')
     hosted_domain = settings.get(SETTINGS_PREFIX + 'hosted_domain')
     signin_advice = settings.get(SETTINGS_PREFIX + 'signin_advice')
     message = request.params.get('message')
+    url = request.params.get('url')
 
-    if 'url' in request.params:
-        query = {'url': request.params['url']}
-        url = request.route_url('auth_signin_redirect', _query=query)
+    if request.authenticated_userid:
+        if url:
+            return HTTPFound(location=url)
+        else:
+            return HTTPFound(location=find_landing_path(request))
+
+    if url:
+        redirect_url = request.route_url('auth_signin_redirect',
+                                         _query={'url': url})
     else:
-        url = request.route_url('auth_signin_redirect')
+        redirect_url = request.route_url('auth_signin_redirect')
 
-    return {'signin_redirect_url': url,
+    return {'signin_redirect_url': redirect_url,
             'message': message,
             'signin_banner': signin_banner,
             'signin_advice': signin_advice,
@@ -64,7 +74,6 @@ def signin_redirect(request):
              permission=NO_PERMISSION_REQUIRED)
 def callback(request):
     settings = request.registry.settings
-    landing_url = settings.get(SETTINGS_PREFIX + 'landing_url', '/')
     max_age = int(settings.get(SETTINGS_PREFIX + 'max_age', 24 * 3600))
 
     try:
@@ -87,7 +96,7 @@ def callback(request):
         state_params = decode_state(request.params['state'])
         url = state_params['url'][0]
     except:
-        url = landing_url
+        url = find_landing_path(request)
 
     headers = remember(request, principal=principal, max_age=max_age)
     return HTTPFound(location=url, headers=headers)